+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talend/design-system
 
+## 3.4.0
+
+### Minor Changes
+
+- 21189557e: feat(design-system): add TagBeta to the Tag familly
+
+## 3.3.0
+
+### Minor Changes
+
+- 57ea150e8: feat(design-system): FloatingDrawer component
+
 ## 3.2.0
 
 ### Minor Changes